broccoli-html-editorオブジェクトを取り出すAPI `px2ce.createBroccoli()` を追加。
